+++ HISTORY.md
@@ -1,6 +1,21 @@
 # Release History
 
 ## RemexHtml x.x.x (not yet released)
+* Further bug-fixes for PHP 8.4 Dom\Document.
+  - It is recommended to set `suppressHtmlNamespace=true` for
+    DOMBuilder if you are using the old `\DOMImplementation`, and set
+    `suppressHtmlNamespace=false` for the new PHP 8.4
+    \Dom\Implementation.
+  - This will be the default in the next major release; in this
+    release the existing default has remained unchanged
+    (`suppressHtmlNamespace` always defaults to `false`), but beware
+    that with `suppressHtmlNamespace=false` and `\DOMImplementation`,
+    tags like `<mw:section>` will not be parsed consistent with the
+    HTML spec, since `\DOMImplementation` makes it impossible to
+    create a tag with the HTML namespace and with a colon in the tag
+    name.  We've elected to preserve the correct prefix (`null`) and
+    local name (`mw:section`) but omit the HTML namespace in this
+    situation.
 
 ## RemexHtml 5.0.0 (2025-06-04)
 * Bug fixes for PHP 8.4 Dom\Document compatibility.